Test dataset, dataviz, and versioning
Test dataset: Removed unit-test creation flag and added a no-parameter command `test` to run the test dataset. Added all test data to a directory. Harmonized dataviz with the test output.
Versioning: Added a flag to the CLI to print the module's version, and the file `.bumpversion.cfg` to properly update the files with the version in them.
cli.py and utils.py: `commands` now directly prints commands to a script and bypasses outputting dry run to a text file. 
Former-commit-id: 6883cc39ef00de6e426ee1f2355acd99db459472